Move Book component into its own folder, Add jsx missing elements
